% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_portfolio_metrics.R
\name{show_portfolio_metrics}
\alias{show_portfolio_metrics}
\title{Visualize important portfolio metrics}
\usage{
show_portfolio_metrics(portfolio_data)
}
\arguments{
\item{portfolio_data}{A data frame containing the different positions of the portfolio and the values for the metrics for each year.
It has to contain the values for at least three years for each position and it has to contain the values for the annual returns for each position.
On top of that it can contain the following metrics as columns:
year_ : obligatory, as already described
annual_return_year_ : obligatory, as already described
volatility : Volatility measures the range of returns of an investment
sharpe_ratio : The Sharpe Ratio measures the risk-adjusted return of a portfolio
max_drawdown : Maximum Drawdown measures the maximum loss from the peak to the trough of a portfolio over a specific period
cagr : CAGR represents the average annual growth rate of an investment over a specific period
alpha : Alpha measures the excess return of a portfolio compared to a benchmark index
beta : Beta measures the sensitivity of a portfolio compared to the overall market. A beta greater than 1 indicates the portfolio is more volatile than the market
It has to be formatted like the ouput of the calculate_portfolio_metrics function from this package}
}
\value{
No values are being returned. Only the plots are shown
}
\description{
This function visualizes important portfolio metrics and should be used together with the calculate_portfolio_metrics function from this package
}
\examples{
portfolio_data <- data.frame(
 position = c("Stock A", "Stock B", "Stock C"),
 year_0 = c(950, 1500, 1350),
 year_1 = c(1000, 1200, 1500),
 year_2 = c(1100, 1500, 2400)
 )
risk_free_rate <- 2
market_returns <- c(3,5)
portfolio_metrics <- calculate_portfolio_metrics(portfolio_data, risk_free_rate, market_returns)

show_portfolio_metrics(portfolio_metrics)
}
